backend/app/core/security.py
import jwt
from fastapi.security import OAuth2PasswordBearer
from passlib.context import CryptContext
from datetime import datetime, timedelta
import uuid

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/token")
pwd_context = CryptContext(schemes=["bcrypt"],deprecated="auto")

ALGORITHM ="HS256"
ACESS_TOKEN_EXPIRE_MINUTES = 60 
SECRET_KEY = "SECRET"
def get_password_hash(paswword:str) ->str:
    return pwd_context.hash(paswword)

def verify_password(plain_password:str, hashed_password:str) -> bool:
    return pwd_context.verify(plain_password,hashed_password)

def create_access_token(*,data:dict,expire_delta: timedelta ):
    to_encode = data.copy()
    if expire_delta:
        expire = datetime.utcnow() + expire_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp":expire})
    encode_jwt = jwt.encode(to_encode,SECRET_KEY,algorithm=ALGORITHM)
    return encode_jwt


def create_token():
    uuidToken = uuid.uuid1()
    return uuidToken



